+++ ep-sdk/src/types.ts
@@ -1,5 +1,0 @@
-export type EpSdkPagination = {
-  count: number;
-  pageNumber: number;
-  nextPage?: number;
-}